requirements.txt
Django>=5.0,<6.0
psycopg2-binary>=2.9
dj-database-url>=2.1
python-decouple>=3.8
